assignment-3.2-multi-agent-system/requirements.txt
openai>=1.0.0
numpy>=1.21.0
pandas>=1.3.0
jupyter>=1.0.0
python-dotenv>=0.19.0
rich>=12.0.0
asyncio-throttle>=1.0.0
tiktoken>=0.4.0
